// src/types.ts
export interface Producto {
  id: number;
  codigo_barras: string;
  nombre: string;
  precio: number;
  costo?: number;
  unidad_medida?: 'UND' | 'KG' | 'L' | 'PQ' | string; // Presentación: Unidad/Pieza, Kilogramo, Litro, Paquete
  exento_iva?: boolean; // true = Exento de IVA (0%), false/undefined = Gravado con IVA (16%)
  categoria?: string;
  descripcion?: string;
  imagen?: string;
  destacadoTienda?: boolean;
}

export interface InventarioItem {
  id: number;
  sucursal_id: number;
  producto_id: number;
  stock: number;
}
// ...
export interface EmpresaConfig {
  nombreEmpresa: string;
  rif: string;
  direccionFiscal: string;
  telefono: string;
  logoUrl?: string;
  tasaCambio: number; // Bolívares por USD (ej. 36.50)
  fechaTasa: string;
  ultimaActualizacionTasa?: string;
  nombreTienda1: string;
  nombreTienda2: string;
  nombreOficina: string;
}